Structuring Index.html
To structure index.html, I divided the body to 4 sections and a footer.
This allows me to edit and style one section at a time. I also linked
the styles.css file on index.html. I started some of the styles
such as the display: flex for the body, and the background in the first
section.

The first section is divided into to containers to separate the header
and the other info under the header.