Finished The First Section
The content of the first section of the landing page is added to the
index.html, and styled in the styles.css.

The first section is divided into two columns, the first one is the
header, the second one is further divided into two rows left and right.
The css file is structured so that there are two flex blocks inside the
header flex block, and then two blocks inside the block under the header.